Added common md component for renderer doc pages.
