--- packages/react-renderer-demo/src/common/doc-texts/data-types.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import MdxContent from 'docs/components/data-types.md';
-
-export default (
-  <React.Fragment>
-    <MdxContent />
-  </React.Fragment>
-);
