--- packages/react-renderer-demo/src/common/doc-texts/form-controls.js
@@ -1,4 +1,0 @@
-import React from 'react';
-import MdxContent from 'docs/components/form-controls.md';
-
-export default <MdxContent />;
