--- packages/react-renderer-demo/src/common/doc-texts/installation.js
@@ -1,4 +1,0 @@
-import React from 'react';
-import MdxContent from 'docs/components/installation.md';
-
-export default (<MdxContent />);
