+++ packages/react-renderer-demo/src/docs-components/get-started.md
@@ -1,3 +1,5 @@
+import RawComponent from '../common/component/raw-component';
+
 ## React form renderer
 
 React form renderer is a component designed for ManageIQ and
@@ -6,3 +8,5 @@ and renders them into react components.
 It uses [React final form](https://github.com/final-form/react-final-form) for the form state management.
 It is highly recommended to check their documentations first to fully understand how
 the [data-driven-forms](https://github.com/data-driven-forms/react-forms) libraries work.
+
+<RawComponent source="get-started/get-started" />
